+++ src/maple-fighters/Assets/Maple Fighters/Scripts/Network/Peer Logics/RegistrationPeerLogic.cs
@@ -15,7 +15,7 @@ namespace Scripts.Services
                 await ServerPeerHandler
                     .SendOperation<RegisterRequestParameters, RegisterResponseParameters>(
                         yield,
-                        RegistrationOperations.Register,
+                        (byte)RegistrationOperations.Register,
                         parameters,
                         MessageSendOptions.DefaultReliable());
         }
